backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
huggingface_hub
PyJWT==2.3.0
cryptography==36.0.1
asyncio
httpx